First draft - Learning Java Challenge project.

+++ src/calculator/Calculation.java
@@ -1,8 +1,6 @@
 package calculator;
 
-import java.util.Objects;
-
 public interface Calculation {
-    int calculation(int input1, int input2);
+    float calculation(float input1, float input2);
 
 }
